Bro tf
I think the v3 is the one that works the best right now. I'm probably just gonna stick to that one or something idk. I have to practice Loops, If/Else more. Arrays are interesting but they're not what controls the code. It stores the information but has not- like, it isn't what's guiding the code from one thing to another in a sense of this or that.