Flushing it out more
The To do section is for things I want the program to do which is subject to change.
The Do section is a technical side and are things I want to get to do. Ya it's a technical side.
I want to get some flowshart or something done though.
Other than that it's all good so far kind of. I want to use functions though. And on the next project I want to start using classes. Then more later I want to link files. I think It's called overload or something. I still kind of know though.
:bath: