Just finished pseudocode
On the first version, I was just trying to make up something with functions that was like tf to me. It's when I called it and stated it twice somewhere I think. And then I'm just like, I'll see when I start typing it.
The second version is when I start to type out most of the code from my pseudocode I think I'm gonna need.
And now it's time to test this first one.
:+1: